refactor: query builder toolbar and status
